backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Configuration
pyyaml>=6.0.1

# HTTP client
httpx>=0.26.0

# Content extraction
crawl4ai>=0.2.0
youtube-transcript-api>=0.6.0
pymupdf>=1.23.0

# LLM (MiniMax is OpenAI-compatible)
openai>=1.12.0

# Server-Sent Events
sse-starlette>=2.0.0

# Utilities
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dateutil>=2.8.0
